+++ src/components/AppWindow/index.tsx
@@ -1,9 +1,9 @@
 import { AppReducerContext } from '@/store/App/AppContext';
-import { RefObject, useContext } from 'react';
+import { RefObject, useContext, useRef, useState } from 'react';
 import CloseIcon from '@/assets/icons/AppWindow/Close.svg?react';
 import MinimizeIcon from '@/assets/icons/AppWindow/Minimize.svg?react';
 import FullScreenIcon from '@/assets/icons/AppWindow/FullScreen.svg?react';
-import useDragElement from '@/hooks/useDragElement';
+import { getPosition, mouseDrag } from '@/utils/mouseDrag';
 
 interface Props {
   title: string;
@@ -17,7 +17,19 @@ const AppWindow = ({ title, id, container, initialX, initialY }: Props) => {
   const dispatch = useContext(AppReducerContext);
   if (!dispatch) throw new Error('dispatch is null');
 
-  const { x, y, targetRef, handleDragElement } = useDragElement<HTMLDivElement>(container, initialX, initialY);
+  const [{ x, y }, setPosition] = useState({ x: initialX, y: initialY });
+  const targetRef = useRef<HTMLDivElement>(null);
+
+  const { handleMouseDown } = mouseDrag((moveX, moveY) => {
+    if (container.current && targetRef.current) {
+      const containerRect = container.current.getBoundingClientRect();
+      const targetElementRect = targetRef.current.getBoundingClientRect();
+      const calculatedX = getPosition(x + moveX, 0, containerRect.width - targetElementRect.width);
+      const calculatedY = getPosition(y + moveY, 0, containerRect.height - targetElementRect.height);
+      setPosition({ x: calculatedX, y: calculatedY });
+    }
+  });
+
   const handleClose = () => dispatch({ type: 'CLOSE', id });
 
   return (
@@ -26,7 +38,7 @@ const AppWindow = ({ title, id, container, initialX, initialY }: Props) => {
       style={{ transform: `translate(${x}px, ${y}px)` }}
       ref={targetRef}
     >
-      <div className='flex h-7 w-full items-center justify-center bg-[#e4e4e4]' onMouseDown={handleDragElement}>
+      <div className='flex h-7 w-full items-center justify-center bg-[#e4e4e4]' onMouseDown={handleMouseDown}>
         <div className='hidden-wrapper absolute left-2 flex items-center gap-1'>
           <button
             className='flex h-3 w-3 cursor-pointer items-center justify-center rounded-full border-[0.5px] border-[#00000033] bg-[#FF5F57]'
